requirements.txt
openai>=1.50.0
pydantic>=2.7.0
PyYAML>=6.0.1
tenacity>=8.2.3
fastapi>=0.112.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
